refactor: cleanup io
Make io modules only contain abstractions of external io logic. Move all other code either into domain (if it is pure) or app (if it is impure)
