+++ src/app/query-registries.ts
