+++ src/cli/cmd-deps.ts
@@ -16,10 +16,10 @@ import { SemanticVersion } from "../domain/semantic-version";
 import { getUserUpmConfigPathFor } from "../domain/upm-config";
 import type { CheckUrlExists } from "../io/check-url";
 import type { GetRegistryPackument } from "../io/packument-io";
-import { getHomePathFromEnv } from "../io/special-paths";
+import { getHomePathFromEnv } from "../domain/special-paths";
 import type { ReadTextFile } from "../io/text-file-io";
 import { DebugLog } from "../domain/logging";
-import { queryAllRegistriesLazy } from "../app/sources";
+import { queryAllRegistriesLazy } from "../app/query-registries";
 import { isZod } from "../domain/zod-utils";
 import { stringifyDependencyGraph } from "./dependency-logging";
 import { CmdOptions } from "./options";
