+++ src/domain/npmrc.ts
@@ -1,5 +1,6 @@
 import path from "path";
 import { RegistryUrl } from "./registry-url";
+import { splitLines } from "./string-utils";
 
 /**
  * The content lines of a npmrc file.
@@ -54,4 +55,24 @@ export function setToken(
   // Remove empty lines
   lines = lines.filter((l) => l);
   return lines;
+}
+
+/**
+ * Parses the content of a `.npmrc` file.
+ * @param content The file content.
+ * @returns The parsed content.
+ */
+export function parseNpmrc(content: string): Npmrc {
+  // TODO: Check if lines are valid.
+  return splitLines(content);
+}
+
+/**
+ * Serializes a {@link Npmrc} object so it can be written to a `.npmrc` file.
+ * @param eol A string to use for line separators.
+ * @param npmrc The npmrc.
+ * @returns The file content.
+ */
+export function serializeNpmrc(eol: string, npmrc: Npmrc): string {
+  return npmrc.join(eol);
 }
