+++ src/domain/project-manifest.ts
@@ -1,12 +1,12 @@
 import path from "node:path";
 import { z } from "zod";
-import { removeRecordKey } from "./record-utils";
-import { removeTrailingSlash } from "./string-utils";
 import { DomainName } from "./domain-name";
 import { PackageUrl } from "./package-url";
+import { removeRecordKey } from "./record-utils";
 import { RegistryUrl } from "./registry-url";
 import { ScopedRegistry } from "./scoped-registry";
 import { SemanticVersion } from "./semantic-version";
+import { removeTrailingSlash } from "./string-utils";
 
 /**
  * The content of the project-manifest (manifest.json) of a Unity project.
@@ -237,4 +237,24 @@ const projectManifestSchema = z.object({}).passthrough();
 export function parseProjectManifest(content: string): UnityProjectManifest {
   const json = JSON.parse(content);
   return projectManifestSchema.parse(json) as UnityProjectManifest;
+}
+
+/**
+ * Serializes a {@link UnityProjectManifest} object into json format.
+ * @param manifest The manifest to serialize.
+ * @returns The serialized manifest.
+ */
+export function serializeProjectManifest(
+  manifest: UnityProjectManifest
+): string {
+  // Remove empty scoped registries
+  if (manifest.scopedRegistries !== undefined)
+    manifest = {
+      ...manifest,
+      scopedRegistries: manifest.scopedRegistries.filter(
+        (it) => it.scopes.length > 0
+      ),
+    };
+
+  return JSON.stringify(manifest, null, 2);
 }
